feat: replace Siderail with Siderail2
