+++ packages/components/src/templates/next/components/internal/Siderail/Siderail.tsx
@@ -1,151 +1,50 @@
 "use client"
 
-import { useState } from "react"
-import { BiLeftArrowAlt } from "react-icons/bi"
-import { MdKeyboardArrowDown, MdKeyboardArrowUp } from "react-icons/md"
+import { useMemo } from "react"
+import { tv } from "tailwind-variants"
 
+import type { Item } from "./types"
 import type { SiderailProps } from "~/interfaces"
+import { SiderailProvider } from "./SiderailContext"
+import { SiderailLabel } from "./SiderailLabel"
+import { SiderailList } from "./SiderailList"
+import { SiderailParentLabel } from "./SiderailParentLabel"
+import { isNestableItem } from "./utils"
 
-const SiderailMobile = ({
-  parentTitle,
-  pages,
-  LinkComponent = "a",
-}: Omit<SiderailProps, "type">) => {
-  const [isExpanded, setIsExpanded] = useState(false)
-  return (
-    <div className="relative lg:hidden">
-      <button
-        className="flex w-full items-center justify-between gap-2 border-b-2 border-black px-5 py-4"
-        onClick={() => setIsExpanded(!isExpanded)}
-        aria-label={`${
-          isExpanded ? "Collapse" : "Expand"
-        } local navigation menu`}
-      >
-        <h4 className="text-content-strong text-heading-04-medium">
-          {parentTitle}
-        </h4>
-        {isExpanded ? (
-          <MdKeyboardArrowUp className="h-auto w-6 flex-shrink-0" />
-        ) : (
-          <MdKeyboardArrowDown className="h-auto w-6 flex-shrink-0" />
-        )}
-      </button>
+const createSiderailStyles = tv({
+  base: "flex flex-col border-b border-b-divider-subtle last:border-b-transparent",
+})
 
-      {isExpanded && (
-        <ul className="absolute w-full bg-white">
-          {pages.map(({ url, title, isCurrent, childPages }) => {
-            return (
-              <li
-                className="border-b border-divider-medium text-content text-paragraph-03"
-                aria-current={isCurrent ? "page" : undefined}
-              >
-                {isCurrent ? (
-                  <p className="block px-5 py-4 font-bold">{title}</p>
-                ) : (
-                  <LinkComponent
-                    href={url}
-                    className="block px-5 py-4 hover:bg-interaction-sub active:underline active:underline-offset-2"
-                  >
-                    {title}
-                  </LinkComponent>
-                )}
-                {isCurrent && (
-                  <ul>
-                    {childPages?.map(({ title, url }, index) => {
-                      return (
-                        <li>
-                          <LinkComponent
-                            href={url}
-                            className={`block py-3 pl-12 pr-5 hover:bg-interaction-sub active:underline active:underline-offset-2 ${
-                              index === childPages.length - 1 && "pb-4"
-                            }`}
-                          >
-                            {title}
-                          </LinkComponent>
-                        </li>
-                      )
-                    })}
-                  </ul>
-                )}
-              </li>
-            )
-          })}
-        </ul>
-      )}
-    </div>
-  )
+// Generate recursive sidebar items if nested
+const generateSiderailItems = (items: Item[]): JSX.Element[] => {
+  const styles = createSiderailStyles()
+  return items.map((item, index) => {
+    if (isNestableItem(item)) {
+      return (
+        <SiderailList key={index} item={item} className={styles}>
+          {generateSiderailItems(item.childPages)}
+        </SiderailList>
+      )
+    }
+    // No children
+    return <SiderailLabel key={index} {...item} className={styles} />
+  })
 }
 
-const SiderailDesktop = ({
+export const Siderail = ({
   parentTitle,
   parentUrl,
   pages,
   LinkComponent = "a",
-}: Omit<SiderailProps, "type">) => {
-  return (
-    <div className="hidden w-full flex-col lg:flex">
-      <LinkComponent
-        href={parentUrl}
-        className="flex items-start gap-2 border-b-2 border-black pb-2"
-      >
-        <BiLeftArrowAlt className="h-9 w-6 flex-shrink-0" />
-        <h4 className="text-content-strong text-heading-04-medium">
-          {parentTitle}
-        </h4>
-      </LinkComponent>
+}: SiderailProps): JSX.Element => {
+  const siderailItems = useMemo(() => generateSiderailItems(pages), [pages])
 
-      <ul>
-        {pages.map(({ url, title, isCurrent, childPages }, index) => {
-          return (
-            <li
-              className={`text-paragraph-03 ${
-                index !== pages.length - 1 && "border-b border-divider-medium"
-              } text-content`}
-              aria-current={isCurrent ? "page" : undefined}
-            >
-              {isCurrent ? (
-                <p className="block px-2 py-3 font-bold">{title}</p>
-              ) : (
-                <LinkComponent
-                  href={url}
-                  className="block px-2 py-3 hover:bg-interaction-sub active:underline active:underline-offset-2"
-                >
-                  <p className="">{title}</p>
-                </LinkComponent>
-              )}
-              {isCurrent && (
-                <ul>
-                  {childPages?.map(({ url, title }, index) => {
-                    return (
-                      <li>
-                        <LinkComponent
-                          href={url}
-                          className={`block py-2.5 pl-10 pr-2 hover:bg-interaction-sub active:underline active:underline-offset-2 ${
-                            index === childPages.length - 1 && "pb-3"
-                          }`}
-                        >
-                          {title}
-                        </LinkComponent>
-                      </li>
-                    )
-                  })}
-                </ul>
-              )}
-            </li>
-          )
-        })}
-      </ul>
+  return (
+    <div className="text-base-content">
+      <SiderailProvider LinkComponent={LinkComponent}>
+        <SiderailParentLabel parentTitle={parentTitle} parentUrl={parentUrl} />
+        <div>{siderailItems}</div>
+      </SiderailProvider>
     </div>
   )
 }
-
-export const Siderail = (props: SiderailProps) => {
-  return (
-    <>
-      <SiderailMobile {...props} />
-      <SiderailDesktop {...props} />
-    </>
-  )
-}
-
-export default Siderail
